Add ability to provide custom cache instance
The desire for implementation details specific to caches comes up regularly (#6, #13). Since Dataloader *uses* a cache rather than implements one, allowing for a choose-your-own-cache allows for the greatest flexibility.

Still, the default will be a simple boundless cache, as most uses of DataLoader are in the context of a single request.