Resolve cache access when values are dispatched
When cached keys are loaded along with new keys, make sure the promises
returned for the cached values are resolved in the same micro-task as
those for the newly loaded values.

This means calling code will not (easily) be able to observe whether a
key was cached or not. If calling code uses the values to perform more
loads (perhaps using different data loaders) this ensures that those
loads are combined into a single batch.

In the following example the same loader is used. Without this commit, once
`userLoader.load(1)` is cached, the subsequent loads unexpectedly result in
three requests, not two:

```js
var DataLoader = require('dataloader')

var userLoader = new DataLoader(keys => myBatchGetUsers(keys));

userLoader.load(1)
  .then(firstUser => {
    // Later…

    userLoader.load(1)
      .then(user => userLoader.load(user.invitedByID))
      .then(invitedBy => console.log(`User 1 was invited by ${invitedBy}`));

    // Elsewhere in your application
    userLoader.load(2)
      .then(user => userLoader.load(user.lastInvitedID))
      .then(lastInvited => console.log(`User 2 last invited ${lastInvited}`));
  })
```

This is because `userLoader.load(1)` resolves nearly instantaneously while
`userLoader.load(2)` requires a backend round-trip. This means the subsequent
loads for `user.invitedByID` and `user.lastInvitedID` are not batched together.

*With* this commit however both promises resolve in the same micro-task,
allowing the subsequent loads to be batched.

Fixes #97.